Add jumbotron
Jumbotron is defined by a div with the jumbotron
class. Content to be centered within the jumbotron
is put in another div with the jumbotron-content
class. The jumbotron is positioned to be behind the
navigation bar and have a parallax scrolling effect
on the background image. In height it takes up the
height of the entire viewport, however, this can be
slimmed down to only half of the viewport by adding
the 'slim' class the the main jumbotron div.