Add content style
The main content for the page is placed just
beneath the jumbotron. If the jumbotron is slim,
the 'slim' class must be added to the 'content'
div as well. The  class can be added to
the main content div in order to make it take
up the full width of the page.