Add footer
Footer is fixed to the bottom of the screen with
the z-index defined on the jumbotron and content
sections causing it to only display when the user
scrolls to the bottom, allowed by an extra margin
added to the content section. The footer also has
a sort of parallax effect as well.